// src/ProductService/proto/product/ProductRequest.ts
// Original file: product.proto


export interface ProductRequest {
  'productName'?: (string);
  'productDescription'?: (string);
  'productPrice'?: (number);
  'productImage'?: (string);
}

export interface ProductRequest__Output {
  'productName': (string);
  'productDescription': (string);
  'productPrice': (number);
  'productImage': (string);
}